Migrate socket native functions to java interoperability
