+++ src/main/ballerina/src/socket/udp_client_socket.bal
@@ -14,6 +14,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
+import ballerinax/java;
+
 # Represents UDP socket client and related remote functions.
 #
 # + localPort - the local port number to which this socket is bound
@@ -33,7 +35,7 @@ public type UdpClient client object {
     # + config - Configure additional details like read timeout etc.
     public function __init(Address? localAddress = (), UdpClientConfig? config = ()) {
         UdpClientConfig configuration = config ?: {};
-        var initResult = self.initEndpoint(localAddress, configuration);
+        var initResult = udpInitEndpoint(self, localAddress, configuration);
         if (initResult is error) {
             panic initResult;
         }
@@ -42,14 +44,14 @@ public type UdpClient client object {
         }
     }
 
-    function initEndpoint(Address? localAddress, UdpClientConfig config) returns error? = external;
-
     # Send given data to the specified remote client.
     #
     # + content - the content that wish to send to the client socket
     # + address - the address of the remote client socket
     # + return - number of bytes got written or an error if encounters an error while writing
-    public remote function sendTo(byte[] content, Address address) returns int|Error = external;
+    public remote function sendTo(byte[] content, Address address) returns int|Error {
+        return udpClientSendTo(self, content, address);
+    }
 
     # Reads data from the remote client. If the data has the specified length, then wait until that number of bytes
     # are received from the client. Else, return the data available in the OS buffer or wait until data receive.
@@ -58,12 +60,16 @@ public type UdpClient client object {
     # + length - Positive integer. Represents the number of bytes which should be read
     # + return - Content as a byte array, the number of bytes read and the address of the sender
     # or an error if encounters an error while reading
-    public remote function receiveFrom(int length = -100) returns [byte[], int, Address]|ReadTimedOutError = external;
+    public remote function receiveFrom(int length = -100) returns [byte[], int, Address]|ReadTimedOutError {
+        return externReceiveFrom(self, length);
+    }
 
     # Closes the client socket connection.
     #
     # + return - - an error if encounters an error while closing the connection or returns nil otherwise
-    public remote function close() returns Error? = external;
+    public remote function close() returns Error? {
+        return udpClientClose(self);
+    }
 };
 
 # This represent the IP socket address.
@@ -81,3 +87,27 @@ public type Address record {|
 public type UdpClientConfig record {|
     int readTimeoutInMillis = 300000;
 |};
+
+function udpInitEndpoint(UdpClient udpClient, Address? localAddress, UdpClientConfig config) returns error? =
+@java:Method {
+    name: "initEndpoint",
+    class: "org.ballerinalang.stdlib.socket.endpoint.udp.client.Utils"
+} external;
+
+function udpClientClose(UdpClient udpClient) returns Error? =
+@java:Method {
+    name: "close",
+    class: "org.ballerinalang.stdlib.socket.endpoint.udp.client.Utils"
+} external;
+
+function externReceiveFrom(UdpClient udpClient, int length) returns [byte[], int, Address]|ReadTimedOutError =
+@java:Method {
+    name: "receiveFrom",
+    class: "org.ballerinalang.stdlib.socket.endpoint.udp.client.Utils"
+} external;
+
+function udpClientSendTo(UdpClient udpClient, byte[] content, Address address) returns int|Error =
+@java:Method {
+    name: "sendTo",
+    class: "org.ballerinalang.stdlib.socket.endpoint.udp.client.Utils"
+} external;
