--- src/main/java/org/ballerinalang/stdlib/socket/endpoint/tcp/client/InitEndpoint.java
@@ -1,81 +1,0 @@
-/*
- * Copyright (c) 2018 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.ballerinalang.stdlib.socket.endpoint.tcp.client;
-
-import org.ballerinalang.jvm.scheduling.Strand;
-import org.ballerinalang.jvm.values.MapValue;
-import org.ballerinalang.jvm.values.ObjectValue;
-import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.natives.annotations.BallerinaFunction;
-import org.ballerinalang.natives.annotations.Receiver;
-import org.ballerinalang.stdlib.socket.tcp.SocketService;
-import org.ballerinalang.stdlib.socket.tcp.SocketUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.net.SocketException;
-import java.nio.channels.SocketChannel;
-
-import static org.ballerinalang.stdlib.socket.SocketConstants.CLIENT;
-import static org.ballerinalang.stdlib.socket.SocketConstants.CLIENT_CONFIG;
-import static org.ballerinalang.stdlib.socket.SocketConstants.CLIENT_SERVICE_CONFIG;
-import static org.ballerinalang.stdlib.socket.SocketConstants.IS_CLIENT;
-import static org.ballerinalang.stdlib.socket.SocketConstants.READ_TIMEOUT;
-import static org.ballerinalang.stdlib.socket.SocketConstants.SOCKET_KEY;
-import static org.ballerinalang.stdlib.socket.SocketConstants.SOCKET_PACKAGE;
-import static org.ballerinalang.stdlib.socket.SocketConstants.SOCKET_SERVICE;
-
-/**
- * Initialize the client socket endpoint.
- *
- * @since 0.985.0
- */
-@BallerinaFunction(
-        orgName = "ballerina",
-        packageName = "socket",
-        functionName = "initEndpoint",
-        receiver = @Receiver(type = TypeKind.OBJECT, structType = CLIENT, structPackage = SOCKET_PACKAGE),
-        isPublic = true
-)
-public class InitEndpoint {
-    private static final Logger log = LoggerFactory.getLogger(InitEndpoint.class);
-
-    public static Object initEndpoint(Strand strand, ObjectValue client, MapValue<String, Object> config) {
-        Object returnValue = null;
-        try {
-            SocketChannel socketChannel = SocketChannel.open();
-            socketChannel.configureBlocking(true);
-            socketChannel.socket().setReuseAddress(true);
-            client.addNativeData(SOCKET_KEY, socketChannel);
-            client.addNativeData(IS_CLIENT, true);
-            ObjectValue callbackService = (ObjectValue) config.get(CLIENT_SERVICE_CONFIG);
-            client.addNativeData(CLIENT_CONFIG, config);
-            final long readTimeout = config.getIntValue(READ_TIMEOUT);
-            client.addNativeData(SOCKET_SERVICE,
-                    new SocketService(socketChannel, strand.scheduler, callbackService, readTimeout));
-        } catch (SocketException e) {
-            returnValue = SocketUtils.createSocketError("unable to bind the local socket port");
-        } catch (IOException e) {
-            log.error("Unable to initiate the client socket", e);
-            returnValue = SocketUtils.createSocketError("unable to initiate the socket");
-        }
-        return returnValue;
-    }
-}
