--- src/main/java/org/ballerinalang/stdlib/socket/endpoint/tcp/client/Read.java
@@ -1,75 +1,0 @@
-/*
- * Copyright (c) 2019 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.ballerinalang.stdlib.socket.endpoint.tcp.client;
-
-import org.ballerinalang.jvm.scheduling.Strand;
-import org.ballerinalang.jvm.values.ObjectValue;
-import org.ballerinalang.jvm.values.connector.NonBlockingCallback;
-import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.natives.annotations.BallerinaFunction;
-import org.ballerinalang.natives.annotations.Receiver;
-import org.ballerinalang.stdlib.socket.SocketConstants;
-import org.ballerinalang.stdlib.socket.tcp.ReadPendingCallback;
-import org.ballerinalang.stdlib.socket.tcp.ReadPendingSocketMap;
-import org.ballerinalang.stdlib.socket.tcp.SelectorManager;
-import org.ballerinalang.stdlib.socket.tcp.SocketService;
-import org.ballerinalang.stdlib.socket.tcp.SocketUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.nio.channels.SocketChannel;
-
-import static org.ballerinalang.stdlib.socket.SocketConstants.CLIENT;
-import static org.ballerinalang.stdlib.socket.SocketConstants.DEFAULT_EXPECTED_READ_LENGTH;
-import static org.ballerinalang.stdlib.socket.SocketConstants.SOCKET_PACKAGE;
-
-/**
- * 'Read' method implementation of the socket caller action.
- *
- * @since 0.995.0
- */
-@BallerinaFunction(
-        orgName = "ballerina",
-        packageName = "socket",
-        functionName = "read",
-        receiver = @Receiver(type = TypeKind.OBJECT, structType = CLIENT, structPackage = SOCKET_PACKAGE),
-        isPublic = true
-)
-public class Read {
-    private static final Logger log = LoggerFactory.getLogger(Read.class);
-
-    public static Object read(Strand strand, ObjectValue client, long length) {
-        final NonBlockingCallback callback = new NonBlockingCallback(strand);
-        if (length != DEFAULT_EXPECTED_READ_LENGTH && length < 1) {
-            String msg = "requested byte length need to be 1 or more";
-            callback.setReturnValues(SocketUtils.createSocketError(msg));
-            callback.notifySuccess();
-            return null;
-        }
-        SocketService socketService = (SocketService) client.getNativeData(SocketConstants.SOCKET_SERVICE);
-        SocketChannel socketChannel = (SocketChannel) client.getNativeData(SocketConstants.SOCKET_KEY);
-        int socketHash = socketChannel.hashCode();
-        ReadPendingCallback readPendingCallback = new ReadPendingCallback(callback, (int) length, socketHash,
-                socketService.getReadTimeout());
-        ReadPendingSocketMap.getInstance().add(socketChannel.hashCode(), readPendingCallback);
-        log.debug("Notify to invokeRead");
-        SelectorManager.getInstance().invokeRead(socketHash, socketService.getService() != null);
-        return null;
-    }
-}
