--- src/main/java/org/ballerinalang/stdlib/socket/endpoint/tcp/server/Register.java
@@ -1,61 +1,0 @@
-/*
- * Copyright (c) 2018 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.ballerinalang.stdlib.socket.endpoint.tcp.server;
-
-import org.ballerinalang.jvm.scheduling.Scheduler;
-import org.ballerinalang.jvm.scheduling.Strand;
-import org.ballerinalang.jvm.values.ObjectValue;
-import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.natives.annotations.BallerinaFunction;
-import org.ballerinalang.natives.annotations.Receiver;
-import org.ballerinalang.stdlib.socket.tcp.SocketService;
-
-import java.nio.channels.ServerSocketChannel;
-
-import static org.ballerinalang.stdlib.socket.SocketConstants.READ_TIMEOUT;
-import static org.ballerinalang.stdlib.socket.SocketConstants.SERVER_SOCKET_KEY;
-import static org.ballerinalang.stdlib.socket.SocketConstants.SOCKET_PACKAGE;
-import static org.ballerinalang.stdlib.socket.SocketConstants.SOCKET_SERVICE;
-
-/**
- * Register socket listener service.
- *
- * @since 0.985.0
- */
-@BallerinaFunction(
-        orgName = "ballerina",
-        packageName = "socket",
-        functionName = "register",
-        receiver = @Receiver(type = TypeKind.OBJECT, structType = "Listener", structPackage = SOCKET_PACKAGE),
-        isPublic = true
-)
-public class Register {
-
-    public static Object register(Strand strand, ObjectValue listener, ObjectValue service, Object name) {
-        final SocketService socketService = getSocketService(listener, strand.scheduler, service);
-        listener.addNativeData(SOCKET_SERVICE, socketService);
-        return null;
-    }
-
-    private static SocketService getSocketService(ObjectValue listener, Scheduler scheduler, ObjectValue service) {
-        ServerSocketChannel serverSocket = (ServerSocketChannel) listener.getNativeData(SERVER_SOCKET_KEY);
-        long timeout = (long) listener.getNativeData(READ_TIMEOUT);
-        return new SocketService(serverSocket, scheduler, service, timeout);
-    }
-}
