--- src/main/java/org/ballerinalang/stdlib/socket/endpoint/udp/client/SendTo.java
@@ -1,80 +1,0 @@
-/*
- * Copyright (c) 2019 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.ballerinalang.stdlib.socket.endpoint.udp.client;
-
-import org.ballerinalang.jvm.scheduling.Strand;
-import org.ballerinalang.jvm.values.ArrayValue;
-import org.ballerinalang.jvm.values.MapValue;
-import org.ballerinalang.jvm.values.ObjectValue;
-import org.ballerinalang.model.types.TypeKind;
-import org.ballerinalang.natives.annotations.BallerinaFunction;
-import org.ballerinalang.natives.annotations.Receiver;
-import org.ballerinalang.stdlib.socket.SocketConstants;
-import org.ballerinalang.stdlib.socket.tcp.SocketUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-import java.net.InetSocketAddress;
-import java.nio.ByteBuffer;
-import java.nio.channels.ClosedChannelException;
-import java.nio.channels.DatagramChannel;
-
-import static org.ballerinalang.stdlib.socket.SocketConstants.SOCKET_PACKAGE;
-import static org.ballerinalang.stdlib.socket.SocketConstants.UDP_CLIENT;
-
-/**
- * 'SendTo' method implementation of the UDP socket client action.
- *
- * @since 0.995.0
- */
-@BallerinaFunction(
-        orgName = "ballerina",
-        packageName = "socket",
-        functionName = "sendTo",
-        receiver = @Receiver(type = TypeKind.OBJECT, structType = UDP_CLIENT, structPackage = SOCKET_PACKAGE),
-        isPublic = true
-)
-public class SendTo {
-    private static final Logger log = LoggerFactory.getLogger(SendTo.class);
-
-    public static Object sendTo(Strand strand, ObjectValue client, ArrayValue content,
-            MapValue<String, Object> address) {
-        DatagramChannel socket = (DatagramChannel) client.getNativeData(SocketConstants.SOCKET_KEY);
-        String host = address.getStringValue(SocketConstants.CONFIG_FIELD_HOST);
-        int port = address.getIntValue(SocketConstants.CONFIG_FIELD_PORT).intValue();
-        byte[] byteContent = content.getBytes();
-        if (log.isDebugEnabled()) {
-            log.debug(String.format("No of byte going to write[%d]: %d", socket.hashCode(), byteContent.length));
-        }
-        try {
-            final InetSocketAddress remote = new InetSocketAddress(host, port);
-            int write = socket.send(ByteBuffer.wrap(byteContent), remote);
-            if (log.isDebugEnabled()) {
-                log.debug(String.format("No of byte written for the client[%d]: %d", socket.hashCode(), write));
-            }
-            return write;
-        } catch (ClosedChannelException e) {
-            return SocketUtils.createSocketError("client socket close already.");
-        } catch (IOException e) {
-            log.error("Unable to perform write[" + socket.hashCode() + "]", e);
-            return SocketUtils.createSocketError("write failed. " + e.getMessage());
-        }
-    }
-}
