+++ src/main/java/org/ballerinalang/stdlib/socket/tcp/ReadPendingSocketMap.java
@@ -22,7 +22,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * This map will hold all the pending read action that coming through
- * {@link org.ballerinalang.stdlib.socket.endpoint.tcp.client.Read} action until new content ready.
+ * {@link org.ballerinalang.stdlib.socket.endpoint.tcp.ClientUtils} action until new content ready.
  *
  * @since 0.995.0
  */
